backend/pratham/requirements.txt
fastapi
uvicorn[standard]
sentencepiece
python-docx
langdetect
pymupdf
requests
python-multipart
